requirements.txt
pandas
kafka-python
apache-flink
google-cloud-bigquery
dbt
apache-airflow
flask
